+++ packages/demo/src/App.svelte
@@ -1,20 +1,23 @@
 <script lang="ts">
   import '../../lib/src/index.css';
   import AsyncTableDemo from './AsyncTableDemo.svelte';
-  import Demos from './Demos.svelte';
+ //import Demos from './Demos.svelte';
   import FileListDemo from './FileListDemo.svelte';
   import InputDemo from './InputDemo.svelte';
   import MultiSelectDemo from './MultiSelectDemo.svelte';
   import ImageGalleryDemo from './ImageGalleryDemo.svelte';
   import CkEditorDemo from './CkEditorDemo.svelte';
+  import RangeDemo from './RangeDemo.svelte';
 
 </script>
 
-<AsyncTableDemo />
+<!--<AsyncTableDemo /> -->
 <!-- <MultiSelectDemo /> -->
 <!-- <FileListDemo /> -->
-<!-- <Demos /> -->
+ <!--<Demos />-->
 <!-- <InputDemo /> -->
 <!-- <AsyncTableDemo /> -->
 <!-- <ImageGalleryDemo /> -->
  <!-- <CkEditorDemo /> -->
+
+ <RangeDemo />
