+++ packages/demo/src/RangeDemo.svelte
@@ -1,0 +1,6 @@
+<script>
+  import '../../../dist/range.wc';
+  import '../../../dist/range.css';
+</script>
+
+<jp-range></jp-range>
