+++ packages/lib/src/form-fields/range/range.wc.pcss
@@ -1,0 +1,106 @@
+
+  .slider-container {
+    width: 100%;
+    text-align: left; 
+    margin-bottom: 15px; 
+  }
+
+  
+  .slider {
+    position: relative;
+    min-width: 300px;
+    width: 1px;
+    height: 4px;
+    background-color: #ccc;
+    border-radius: 4px;
+  }
+
+  .progress {
+    position: absolute;
+    background-color: var(--primary-color);
+    top: 0;
+    height: 100%;
+  }
+
+  .range-input {
+    position: absolute;
+    top: 0;
+  }
+
+  input[type='range'] {
+    -webkit-appearance: none;
+    -moz-appearance: none;
+    cursor: pointer;
+    pointer-events: none;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+    padding: 0;
+    border-radius: 4px;
+    outline: none;
+    border: none;
+    background: transparent;
+    accent-color: var(--primary-color);
+  }
+
+  /*Mozilla thumb*/
+  input[type='range']::-moz-range-track {
+    width: 100%;
+    height: 100%;
+    border: none;
+    border-radius: 4px;
+  }
+
+  input[type='range']::-moz-range-thumb {
+    pointer-events: auto;
+    border: none;
+    height: 16px;
+    width: 16px;
+    border-radius: 50%;
+    background: var(--primary-color);
+    transition: 0.2s ease-in-out;
+  }
+
+  input[type='range']:focus::-moz-range-track {
+    background: transparent;
+  }
+
+  input[type='range']::-moz-range-progress {
+    appearance: none;
+    background: transparent;
+    height: 100%;
+    transition-delay: 50ms;
+  }
+  /* Chrome, safari, opera edge */
+  .range-input::-webkit-slider-thumb {
+    pointer-events: auto;
+    -webkit-appearance: none;
+    border: none;
+    height: 16px;
+    width: 16px;
+    border-radius: 50%;
+    background: var(--primary-color);
+    transition: 0.2s ease-in-out;
+  }
+
+  .range-input::-webkit-slider-thumb:hover {
+    box-shadow: 0 0 0 10px rgba(255, 85, 0, 0.1);
+  }
+
+  .range-input:active::-webkit-slider-thumb {
+    box-shadow: 0 0 0 13px rgba(255, 85, 0, 0.281);
+  }
+
+  .range-input:disabled::-webkit-slider-thumb:hover {
+    box-shadow: 0 0 0 10px rgba(131, 128, 127, 0.267);
+  }
+
+  .range-input:disabled:active::-webkit-slider-thumb {
+    box-shadow: 0 0 0 10px rgba(131, 128, 127, 0.267);
+  }
+
+  input[type='range']::-webkit-slider-runnable-track {
+    box-shadow: none;
+    border: none;
+    background: transparent;
+  }
